// src/Chat/types.ts
export type unreadCountType = {
  unreadMessages: number;
  userId: string;
};

export interface selectedChatType {
  name: string;
  chatId: string;
  createdAt: Date;
  email: string;
  id: string;
  lastMessage: string;
  lastMessageCreatedAt: string;
  lastMessageForReceiver:string,
  lastMessageForSender:string,
  senderId:string,
  receiverId:string,
  profileUrl: string;
  refreshToken: String;
  tokenExpiresIn: Date | null;
  unreadCount: unreadCountType;
  publickey:string
}
// ...
 export type MessageStatus   = "pending"|"sent" | "delivered" | "seen"
 export type MessageType = {
  id?: string;
  senderId: string;
  receiverId: string;
  senderContent:string,
  receiverContent:string,
  createdAt: number;
  tempId: string;
  isMedia?: boolean;
  uploading?: boolean;
  chatId: string | null;
  status:MessageStatus;
  error:boolean;
  errorMessage:string | null;
};